backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
requests
beautifulsoup4
python-dotenv
typing-extensions
